files: move files to corresponding dir
