+++ frontend/src/components/Navbar/Navbar.css
@@ -1,0 +1,9 @@
+/* header, nav{
+  width: 100%;
+  height: 15vh;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  background-color: white;
+} */
