+++ frontend/src/components/Navbar/Navbar.jsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import { Link as LinkRouter } from 'react-router-dom';
-import '../styles/navbar.css';
+import './Navbar.css';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {  faBell, faCartShopping} from '@fortawesome/free-solid-svg-icons'
+import HamburgerBtn from '../Buttons/HamburgerBtn/HamburgerBtn';
 
 const pages = [
   {
@@ -21,6 +22,7 @@ function Navbar() {
   return (
     <>
       <header>
+        <HamburgerBtn />
         <nav >
           <ul style={{width:'90%', display:'flex', flexDirection:'row', justifyContent:'flex-end', columnGap:'20px', marginRight:'50px'}}>
             {pages.map((page) => (
