+++ frontend/src/pages/Logo.jsx
@@ -1,0 +1,7 @@
+export default function Welcome() {
+  return (
+    <>
+      <p>Logo</p>
+    </>
+  )
+}
