--- frontend/src/styles/navbar.css
@@ -1,9 +1,0 @@
-header, nav{
-    width: 100%;
-    height: 15vh;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    background-color: white;
-}
